+++ drizzle/0004_absent_colleen_wing.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "month_actual_expense" ADD COLUMN "converted_line_kind" "line_kind";